code structure fix
