+++ .gitignore
@@ -11,6 +11,11 @@
 # production
 /build
 
+#Logs
+package-lock.json
+
+node_modules
+
 # misc
 .DS_Store
 .env.local
