+++ src/index.css
@@ -1,4 +1,3 @@
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -9,25 +8,25 @@ body {
   font-family: "Inter", sans-serif;
 }
 
-.main-container{
+.main-container {
   background-color: #292424;
 }
 .subtext,
-.women{
-  color: #f56630;
+.women {
+  color: #808000;
 }
-.nav-logo{
+.nav-logo {
   width: 1.5rem;
   height: 1.5rem;
 }
-.side-container{
-  background-color:#FEF5F0;
+.side-container {
+  background-color: #fef5f0;
 }
-.first-side{
+.first-side {
   background-color: white;
 }
 
-.first-side .items{
+.first-side .items {
   display: flex;
   justify-content: space-between;
   align-items: center;
@@ -35,6 +34,6 @@ body {
   border-bottom: 2px solid #f0f2f5;
   margin-bottom: 1rem;
 }
-.search{
+.search {
   border: 2px solid red;
 }
