--- src/pages/SideContainer.jsx
@@ -1,7 +1,0 @@
-// import React from "react";
-
-// export default function SideContainer() {
-//   return (
-    
-//   );
-// }
