+++ src/pages/checkout/CheckOut.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CheckOut = () => {
+  return <div>CheckOut</div>;
+};
+
+export default CheckOut;
