+++ src/pages/home/Home.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import AppLayout from "../../components/layout/AppLayout";
+import Body from "../../components/body/Body";
+
+const Home = () => {
+  return (
+    <AppLayout>
+      <Body />
+    </AppLayout>
+  );
+};
+
+export default Home;
